+++ src/fast_jsonl/__init__.py
@@ -1,7 +1,21 @@
+# Copyright 2024 Mathew Huerta-Enochian
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import fast_jsonl.cache as cache
 import fast_jsonl.constants as constants
 import fast_jsonl.lock as lock
-from fast_jsonl.reader import Reader
+from fast_jsonl.reader import Reader, MultiReader
 
 
 __version__ = "0.1.0"