+++ CHANGELOG.md
@@ -6,7 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.1.0] - 2024-07-08
 ### Added
-- This changelog file
+- Initial release of fast-jsonl library:
+  - Source code at `src/fast_jsonl/`.
+  - Project and repository files (`pyproject.toml`, `.gitattributes`)
+  - This changelog file (`CHANGELOG.md`).
+  - A citation file (`CITATION.cff`).
+  - Apache 2.0 license file (`LICENSE`).
+  - Sphinx-based documentation files at `docs/`.
+  - Pytest-based test files at `tests/`.
+  - A GitHub workflow yaml file to run pytests
+    (`.github/workflows/pytest.yml`)
 
 [Unreleased]: https://github.com/mathewhuen/fast-jsonl
+[0.1.0]: https://github.com/mathewhuen/fast-jsonl/releases/tag/v0.1.0